+ authorization and redirections

+++ app/Http/Controllers/VisitsController.php
@@ -12,64 +12,85 @@ class VisitsController
 {
     public function registerVisit($patientId)
     {
+        $user = Auth::user();
 
-        $patient = DB::table('patients')->where('id', $patientId)->first();
-        $doctors = DB::table('doctors')->get();
+        if ($user->role == '2') {
+            $patient = DB::table('patients')->where('id', $patientId)->first();
+            $doctors = DB::table('doctors')->get();
 
-        return view('pages/registerVisit', compact('patient', 'doctors'));
+            return view('pages/registerVisit', compact('patient', 'doctors'));
+        } else {
+            return redirect('/mainPage');
+        }
     }
 
 
     public function postRegisterVisit($patientId, Request $request)
     {
 
-        $patient = DB::table('patients')->where('id', $patientId)->first();
-        $doctors = DB::table('doctors')->get();
+        $user = Auth::user();
 
-        $data = $request->all();
+        if ($user->role == '2') {
+            $patient = DB::table('patients')->where('id', $patientId)->first();
+            $doctors = DB::table('doctors')->get();
 
-        $visit = new Pending_visits();
+            $data = $request->all();
 
-        $visit->date_of_visit = $data['datepicker'];
-        $visit->hour_of_visit = $data['timepicker'];
-        $visit->doctor_id = $data['doctor'];
-        $visit->patient_id = $patientId;
-        $visit->type_visit = $data['type'];
+            $visit = new Pending_visits();
 
-        $visit->save();
-        return view('pages/registerVisit', compact('patient', 'doctors'));
+            $visit->date_of_visit = $data['datepicker'];
+            $visit->hour_of_visit = $data['timepicker'];
+            $visit->doctor_id = $data['doctor'];
+            $visit->patient_id = $patientId;
+            $visit->type_visit = $data['type'];
+
+            $visit->save();
+            return view('pages/registerVisit', compact('patient', 'doctors'));
+        } else {
+            return redirect('/mainPage');
+        }
+
+
     }
 
     public function pendingVisits($patientId)
     {
-        $doctors = DB::table('doctors')->get();
-        $patient = DB::table('patients')->where('id', $patientId)->first();
-        $visits = DB::table('pending_visits')->where('patient_Id', $patientId)
-            ->join('doctors', 'pending_visits.doctor_id', '=', 'doctors.id')
-            ->select('pending_visits.*', 'doctors.first_name', 'doctors.last_name')->get();
-        return view('pages/pendingVisits', compact('visits', 'patient', 'doctors'));
+        $user = Auth::user();
 
+        if ($user->role == '2') {
+            $doctors = DB::table('doctors')->get();
+            $patient = DB::table('patients')->where('id', $patientId)->first();
+            $visits = DB::table('pending_visits')->where('patient_Id', $patientId)
+                ->join('doctors', 'pending_visits.doctor_id', '=', 'doctors.id')
+                ->select('pending_visits.*', 'doctors.first_name', 'doctors.last_name')->get();
+            return view('pages/pendingVisits', compact('visits', 'patient', 'doctors'));
+        } else {
+            return redirect('/mainPage');
+        }
     }
 
     public function removePendingVisit($visitId)
     {
+        $user = Auth::user();
 
+        if ($user->role == '2') {
+            $visit = Pending_visits::find($visitId);
+            $patient = Patient::find($visit->patient_id);
+            $visit->delete();
+            $doctors = DB::table('doctors')->get();
+            $visits = DB::table('pending_visits')->where('patient_Id', $patient->id)
+                ->join('doctors', 'pending_visits.doctor_id', '=', 'doctors.id')
+                ->select('pending_visits.*', 'doctors.first_name', 'doctors.last_name')->get();
+            return view('pages/pendingVisits', compact('visits', 'patient', 'doctors'));
+        } else {
+            return redirect('/mainPage');
+        }
 
-        $visit = Pending_visits::find($visitId);
-        $patient = Patient::find($visit->patient_id);
-        $visit->delete();
-        $doctors = DB::table('doctors')->get();
-        $visits = DB::table('pending_visits')->where('patient_Id', $patient->id)
-            ->join('doctors', 'pending_visits.doctor_id', '=', 'doctors.id')
-            ->select('pending_visits.*', 'doctors.first_name', 'doctors.last_name')->get();
-        return view('pages/pendingVisits', compact('visits', 'patient', 'doctors'));
 
     }
 
-
     public function ajaxDate(Request $request)
     {
-
         $data = $request->all();
 
         $query = DB::table('pending_visits')->select('hour_of_visit')
